requirements.txt
# Required dependencies

# === Transcription (baseline) ===
torch>=2.0.0
torchaudio>=2.0.0
whisperx>=3.1.0
pyannote.audio>=3.1.0
librosa>=0.10.0
pandas>=2.0.0

# === Post-processing ===
openai>=1.0.0
python-dotenv>=1.0.0

# === Evaluation ===
jiwer>=3.0.0
pyannote.core>=5.0.0
pyannote.metrics>=3.2.0
num2words>=0.5.0
numpy>=1.24.0
scipy>=1.10.0
